Move old data cleaning jobs to scheduler
