+++ backend/functions/src/index.ts
@@ -36,7 +36,6 @@ export * from './scheduled/update-view-embeddings'
 export * from './scheduled/update-league'
 export * from './scheduled/update-league-ranks'
 export * from './scheduled/streak-expiration-notification'
-export * from './scheduled/clean-old-rows'
 export * from './scheduled/denormalize-answers'
 export * from './scheduled/score-groups'
 export * from './scheduled/update-view-embeddings'
