+++ backend/scheduler/src/jobs/index.ts
@@ -5,6 +5,8 @@ import { sendOnboardingNotificationsInternal } from 'shared/onboarding-helpers'
 import { updateContractViews } from 'shared/update-contract-views'
 import { updateUserMetricsCore } from 'shared/update-user-metrics-core'
 import { updateGroupMetricsCore } from 'shared/update-group-metrics-core'
+import { cleanOldFeedRows } from './clean-old-feed-rows'
+import { cleanOldNotifications } from './clean-old-notifications'
 import { truncateIncomingWrites } from './truncate-incoming-writes'
 
 export function createJobs() {
@@ -39,6 +41,16 @@ export function createJobs() {
       'truncate-incoming-writes',
       '0 0 4 * * *', // 4 AM daily
       truncateIncomingWrites
+    ),
+    createJob(
+      'clean-old-feed-rows',
+      '0 0 1 * * *', // 1 AM daily
+      cleanOldFeedRows
+    ),
+    createJob(
+      'clean-old-notifications',
+      '0 0 2 * * *', // 2 AM daily
+      cleanOldNotifications
     ),
   ]
 }
